% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2p.R
\name{n2p}
\alias{n2p}
\title{Calculate Proportions}
\usage{
n2p(n)
}
\arguments{
\item{n}{a vector of counts.}
}
\value{
\code{n2p} returns a vector of proportions.
}
\description{
\code{n2p} converts a vector of counts to a vector of proportions.
}
\examples{
n2p(c(61, 477, 836, 1007))
}
