% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vno-data.R
\docType{data}
\name{vno}
\alias{vno}
\title{Results of Vietnamese National Olympiads 2010-2020}
\format{
A data frame with 24151 rows and 5 variables:
\describe{
\item{ID}{student ID}
\item{Year}{year of award}
\item{Team}{administrative contest team that delegated the student}
\item{Subject}{test subject}
\item{Prize}{award achieved}
}
}
\source{
\doi{10.5281/zenodo.3764691}
}
\usage{
vno
}
\description{
A dataset containing the information of
more than 24,000 awarded students over 11 years.
}
\keyword{datasets}
