% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vnomath-data.R
\docType{data}
\name{vnomath}
\alias{vnomath}
\title{Results of Vietnamese Mathematical Olympiad 2010-2020}
\format{
A data frame with 68 rows and 5 variables:
\describe{
\item{Team}{administrative contest team}
\item{First}{number of First prizes}
\item{Second}{number of Second prizes}
\item{Third}{number of Third prizes}
\item{Consolation}{number of Consolation prizes}
}
}
\source{
\doi{10.5281/zenodo.3764691}
}
\usage{
vnomath
}
\description{
A dataset containing the number of prizes in Mathematics over 11 years of
administrative contest teams
}
\keyword{datasets}
